+++ first_ray/bvh.h
@@ -11,10 +11,10 @@ inline int box_x_compare(const void* a, const void* b)
 	hitable* ah = *(hitable * *)a;
 	hitable* bh = *(hitable * *)b;
 
-	if (!ah->bounding_box(0.0f, 0.0f, box_left) || !bh->bounding_box(0.0f, 0.0f, box_right))
+	if (!ah->bounding_box(0.0, 0.0, box_left) || !bh->bounding_box(0.0, 0.0, box_right))
 		std::cerr << "no bounding box in parallel_bvh_node constructor!\n";
 
-	if (box_left.min.x() - box_right.min.x() < 0.0f)
+	if (box_left.min.x() - box_right.min.x() < 0.0)
 		return -1;
 	else
 		return 1;
@@ -28,10 +28,10 @@ inline int box_y_compare(const void* a, const void* b)
 	hitable* ah = *(hitable * *)a;
 	hitable* bh = *(hitable * *)b;
 
-	if (!ah->bounding_box(0.0f, 0.0f, box_left) || !bh->bounding_box(0.0f, 0.0f, box_right))
+	if (!ah->bounding_box(0.0, 0.0, box_left) || !bh->bounding_box(0.0, 0.0, box_right))
 		std::cerr << "no bounding box in parallel_bvh_node constructor!\n";
 
-	if (box_left.min.y() - box_right.min.y() < 0.0f)
+	if (box_left.min.y() - box_right.min.y() < 0.0)
 		return -1;
 	else
 		return 1;
@@ -45,10 +45,10 @@ inline int box_z_compare(const void* a, const void* b)
 	hitable* ah = *(hitable * *)a;
 	hitable* bh = *(hitable * *)b;
 
-	if (!ah->bounding_box(0.0f, 0.0f, box_left) || !bh->bounding_box(0.0f, 0.0f, box_right))
+	if (!ah->bounding_box(0.0, 0.0, box_left) || !bh->bounding_box(0.0, 0.0, box_right))
 		std::cerr << "no bounding box in parallel_bvh_node constructor!\n";
 
-	if (box_left.min.z() - box_right.min.z() < 0.0f)
+	if (box_left.min.z() - box_right.min.z() < 0.0)
 		return -1;
 	else
 		return 1;
@@ -58,22 +58,22 @@ class bvh_node : public hitable
 {
 public:
 	bvh_node() {}
-	bvh_node(hitable **l, int n, float time0, float time1);
-	virtual bool hit(const ray &r, float tmin, float tmax, hit_record &rec) const;
-	virtual bool bounding_box(float t0, float t1, aabb &b) const;
+	bvh_node(hitable **l, int n, double time0, double time1);
+	virtual bool hit(const ray &r, double tmin, double tmax, hit_record &rec) const;
+	virtual bool bounding_box(double t0, double t1, aabb &b) const;
 	
 	hitable *left;
 	hitable *right;
 	aabb box;
 };
 
-bool bvh_node::bounding_box(float t0, float t1, aabb &b) const
+bool bvh_node::bounding_box(double t0, double t1, aabb &b) const
 {
 	b = box;
 	return true;
 }
 
-bool bvh_node::hit(const ray &r, float t_min, float t_max, hit_record &rec) const
+bool bvh_node::hit(const ray &r, double t_min, double t_max, hit_record &rec) const
 {
     // Martin Lamber's BVH improvement https://twitter.com/Peter_shirley/status/1105292977423900673
     if (box.hit(r, t_min, t_max))
@@ -94,11 +94,11 @@ bool bvh_node::hit(const ray &r, float t_min, float t_max, hit_record &rec) cons
 #if 1
 
 // Construct BVH using SAH method
-bvh_node::bvh_node(hitable **l, int n, float time0, float time1)
+bvh_node::bvh_node(hitable **l, int n, double time0, double time1)
 {
 	std::unique_ptr<aabb[]> boxes(new aabb[n]);
-	std::unique_ptr<float[]> left_area(new float[n]);
-	std::unique_ptr<float[]> right_area(new float[n]);
+	std::unique_ptr<double[]> left_area(new double[n]);
+	std::unique_ptr<double[]> right_area(new double[n]);
 	aabb main_box;
 	l[0]->bounding_box(time0, time1, main_box);
 	
@@ -136,11 +136,11 @@ bvh_node::bvh_node(hitable **l, int n, float time0, float time1)
 		right_box = surrounding_box(right_box, boxes[i]);
 		right_area[i] = right_box.area();
 	}
-	float min_SAH = FLT_MAX;
+	double min_SAH = FLT_MAX;
 	int min_SAH_idx;
 	for (int i = 0; i < n - 1; ++i)
 	{
-		float SAH = i*left_area[i] + (n-i -1)*right_area[i+1];
+		double SAH = i*left_area[i] + (n-i -1)*right_area[i+1];
 		if (SAH < min_SAH)
 		{
 			min_SAH_idx = i;
@@ -164,7 +164,7 @@ bvh_node::bvh_node(hitable **l, int n, float time0, float time1)
 }
 
 #else
-bvh_node::bvh_node(hitable **l, int n, float time0, float time1)
+bvh_node::bvh_node(hitable **l, int n, double time0, double time1)
 {
 	int axis = int(3 * gen_cano_rand());
 	if (axis == 0)
